add adlercardbundlestudents and some changes in apis

+++ Application/Features/AdlerCardsUnit/Commands/DeleteAdlerCardsUnitById/DeleteAdlerCardsUnitByIdCommand.cs
@@ -16,14 +16,22 @@ namespace Application.Features
         public class DeleteAdlerCardsUnitByIdCommandHandler : IRequestHandler<DeleteAdlerCardsUnitByIdCommand, Response<int>>
         {
             private readonly IAdlerCardsUnitRepositoryAsync _adlercardsunitRepository;
-            public DeleteAdlerCardsUnitByIdCommandHandler(IAdlerCardsUnitRepositoryAsync adlercardsunitRepository)
+            private readonly IAdlerCardRepositoryAsync _adlercardRepository;
+            public DeleteAdlerCardsUnitByIdCommandHandler(IAdlerCardsUnitRepositoryAsync adlercardsunitRepository,
+                IAdlerCardRepositoryAsync adlercardRepository)
             {
                 _adlercardsunitRepository = adlercardsunitRepository;
+                _adlercardRepository = adlercardRepository;
             }
             public async Task<Response<int>> Handle(DeleteAdlerCardsUnitByIdCommand command, CancellationToken cancellationToken)
             {
                 var adlercardsunit = await _adlercardsunitRepository.GetByIdAsync(command.Id);
                 if (adlercardsunit == null) throw new ApiException($"AdlerCardsUnit Not Found.");
+                var adlerCards = _adlercardRepository.GetAllByUnitId(command.Id);
+                if(adlerCards != null && adlerCards.Count >0)
+                {
+                    throw new ApiException($"Can't delete this unit because it contains alder cards ");
+                }
                 await _adlercardsunitRepository.DeleteAsync(adlercardsunit);
                 return new Response<int>(adlercardsunit.Id);
             }
